+++ pkg/deployment/resources/pod_creator_sync.go
@@ -89,10 +89,16 @@ func (a *ArangoSyncContainer) GetName() string {
 }
 
 func (a *ArangoSyncContainer) GetPorts() []core.ContainerPort {
+	port := shared.ArangoSyncMasterPort
+
+	if a.group == api.ServerGroupSyncWorkers {
+		port = shared.ArangoSyncWorkerPort
+	}
+
 	return []core.ContainerPort{
 		{
 			Name:          shared.ServerContainerName,
-			ContainerPort: int32(shared.ArangoPort),
+			ContainerPort: int32(port),
 			Protocol:      core.ProtocolTCP,
 		},
 	}
@@ -383,6 +389,8 @@ func (m *MemberSyncPod) ApplyPodSpec(spec *core.PodSpec) error {
 	if s := m.groupSpec.SchedulerName; s != nil {
 		spec.SchedulerName = *s
 	}
+
+	m.groupSpec.PodModes.Apply(spec)
 
 	return nil
 }