+++ CHANGELOG.md
@@ -30,6 +30,10 @@
 - (Improvement) Add Resource kerror Type
 - (Bugfix) Do not block reconciliation in case of Resource failure
 - (Improvement) Multi-arch support for ID member
+- (Feature) Allow to change Pod Network and PID settings
+- (Feature) Pre OOM Abort function
+- (Bugfix) Fix ErrorArray String function
+- (Feature) Switch services to Port names
 
 ## [1.2.20](https://github.com/arangodb/kube-arangodb/tree/1.2.20) (2022-10-25)
 - (Feature) Add action progress